NUTRICOACH/requirements.txt
customtkinter
groq
python-dotenv
tkintermapview
geocoder
requests
fpdf2
SpeechRecognition
pyttsx3
PyAudio
